Split attribute models into separate files
