+++ saleor/attribute/models/__init__.py
@@ -1,0 +1,33 @@
+from .base import (
+    Attribute,
+    AttributeTranslation,
+    AttributeValue,
+    AttributeValueTranslation,
+)
+from .page import AssignedPageAttribute, AssignedPageAttributeValue, AttributePage
+from .product import (
+    AssignedProductAttribute,
+    AssignedProductAttributeValue,
+    AttributeProduct,
+)
+from .product_variant import (
+    AssignedVariantAttribute,
+    AssignedVariantAttributeValue,
+    AttributeVariant,
+)
+
+__all__ = [
+    "Attribute",
+    "AttributeTranslation",
+    "AttributeValue",
+    "AttributeValueTranslation",
+    "AssignedPageAttribute",
+    "AssignedPageAttributeValue",
+    "AttributePage",
+    "AssignedProductAttribute",
+    "AssignedProductAttributeValue",
+    "AttributeProduct",
+    "AssignedVariantAttribute",
+    "AssignedVariantAttributeValue",
+    "AttributeVariant",
+]
